Move router logic to timespec-modal component.
