+++ api_server/src/main/frontend/app/eval/time-spec-modal.component.ts
@@ -7,6 +7,8 @@ import { TimeSpecService,
          durationFromString,
          validDurationString }              from './time-spec';
 import { ModalComponent }                   from 'ng2-bs3-modal/ng2-bs3-modal';
+import { Router,
+         ActivatedRoute }                   from '@angular/router';
 
 
 function durationValidator(c: FormControl) {
@@ -50,7 +52,7 @@ export class TimeSpecModalComponent {
   private stepsize: FormControl;
   private formModel: FormGroup;
 
-  constructor(private ts: TimeSpecService, fb: FormBuilder) {
+  constructor(private ts: TimeSpecService, fb: FormBuilder, private _router: Router, private _route: ActivatedRoute) {
     this.begin    = fb.control('', null);  // XXX date validators
     this.end      = fb.control('', null);  // XXX date validators
     this.duration = fb.control('', durationValidator);
@@ -61,16 +63,39 @@ export class TimeSpecModalComponent {
       duration: this.duration,
       stepsize: this.stepsize,
     });
+
+    this._route.queryParams
+        .toPromise()
+        .then((params) => {
+          let p_tsb: string = params['tsb'];
+          let p_tse: string = params['tse'];
+          let p_tsd: string = params['tsd'];
+          let p_tss: string = params['tss'];
+          if (!validDurationString(p_tsd)) p_tsd = null;  // Omit unparsable duration.
+          if (!validDurationString(p_tss)) p_tss = null;  // Omit unparsable stepsize.
+
+          // begin (tsb)
+          let tsb: number = (p_tsb !== null ? parseInt(p_tsb) : null);
+          // end (tse)
+          let tse: number = (p_tse !== null ? parseInt(p_tse) : null);
+          // duration (tsd)
+          let tsd: number = (p_tsd !== null ? durationFromString(p_tsd) : null);
+          // stepsize (tss)
+          let tss: number = (p_tss !== null ? durationFromString(p_tss) : null);
+
+          if (!Number.isFinite(tsb)) tsb = null;  // Omit unparsable number.
+          if (!Number.isFinite(tse)) tse = null;  // Omit unparsable number.
+
+          this.ts.update((tsb != null ? new Date(tsb) : null), (tss != null ? new Date(tse) : null), tsd, tss);
+        });
   }
 
   public open(): void {
     // Initialize form with clean values.
-    this.formModel.value = {
-      begin: dateStr(this.ts.begin),
-      end: dateStr(this.ts.end),
-      duration: durationStr(this.ts.duration),
-      stepsize: durationStr(this.ts.stepsize),
-    };
+    this.begin.setValue(dateStr(this.ts.begin));
+    this.end.setValue(dateStr(this.ts.end));
+    this.duration.setValue(durationStr(this.ts.duration));
+    this.stepsize.setValue(durationStr(this.ts.stepsize));
 
     this.dialog.open();
   }
@@ -80,12 +105,20 @@ export class TimeSpecModalComponent {
   }
 
   private onSubmit() {
-    // Apply update.
+    // Apply update to timespec service.
     this.ts.update(
         decodeDate(this.formModel.value.begin),
         decodeDate(this.formModel.value.end),
         decodeDuration(this.formModel.value.duration),
         decodeDuration(this.formModel.value.stepsize));
+
+    // Apply update to URL query parameters.
+    let params: Map<string, string> = new Map<string, string>();
+    if (this.ts.begin != null)    params['tsb'] = this.ts.begin.getTime().toString();
+    if (this.ts.end != null)      params['tse'] = this.ts.end.getTime().toString();
+    if (this.ts.duration != null) params['tsd'] = durationToString(this.ts.duration);
+    if (this.ts.stepsize != null) params['tss'] = durationToString(this.ts.stepsize);
+    this._router.navigate(this._route.pathFromRoot, { skipLocationChange: true, queryParams: params });
 
     this.dialog.close();
   }
